:truck: VMInfo -> VMSideInfo
